refactor: extract timeline routes into routes/timeline.ts
Pure structural move out of createApp; no handler logic changed. Moves the
timeline views + exports (swimlane, timeline-gaps [+ /hypothesize],
incident-timeline.csv, timeline.jsonl, super-timeline [+ .jsonl, /label,
/promote]). No RouteContext graduations needed — reuses the stable surface
(options, hasAiProvider, serverLogger) and the already-graduated
resynthesizeInBackground. Removed the now-dead toTimesketchJsonlFromList import
from server.ts. dwell-windows (analysisGraph) and activity-log (caseLifecycle)
intentionally left in createApp.
